requirements.txt
streamlit>=1.28.0
pandas>=2.1.1
geopandas>=0.14.0
shapely>=2.0.2
requests>=2.31.0
numpy>=1.26.0
xarray>=2023.9.0
netCDF4>=1.6.4
cdsapi>=0.6.1
rasterio>=1.3.9
pyproj>=3.6.1
folium>=0.14.0
streamlit-folium>=0.15.0
openpyxl>=3.1.2
fiona>=1.9.5
python-dotenv>=1.0.0
earthengine-api>=0.1.384
